products/JGPT/JGPTa/jgpt-api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jgpt-api"
version = "0.3.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic[email]>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9.9",
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.5",
  "pypdf>=4.2.0",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.9",
  "pytest>=8.0.0",
  "requests>=2.32.0",
  "readability-lxml>=0.8.1",
  "markdownify>=0.12.1",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "Pillow>=10.2.0",
  "PyMuPDF>=1.23.25",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<4.0.0",
  "python-jose[cryptography]>=3.3.0",
  "pytesseract>=0.3.10",
  "pdf2image>=1.17.0",
  "pinecone-client>=3.0.0",
  "slowapi>=0.1.9",
  "pymysql>=1.1.0",
  "pymssql>=2.2.11",
  "cryptography>=42.0.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "APScheduler>=3.10.0",
]

[tool.setuptools.packages.find]
include = ["app*", "eval*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]
